Rename 'routes' to router
That's really what this is: a router middleware.
