+++ test/server.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import path from 'path';
 import favicon from 'serve-favicon';
 
-import avatarsRoutes from '../src/routes';
+import avatarsRouter from '../src/router';
 
 const app = express();
 const port = Number(process.env.PORT) || 3002;
@@ -10,7 +10,7 @@ const port = Number(process.env.PORT) || 3002;
 const faviconPath = path.join(__dirname, 'favicon.ico');
 app.use(favicon(faviconPath));
 
-app.use('/avatars', avatarsRoutes);
+app.use('/avatars', avatarsRouter);
 
 app.listen(port, () =>
   console.log(`[Adorable Avatars] Running at: http://localhost:${port}`),
